+++ approvalcrest-testng/src/main/java/com/github/karsaig/approvalcrest/testng/TestNgTestMeta.java
@@ -2,7 +2,6 @@ package com.github.karsaig.approvalcrest.testng;
 
 import com.github.karsaig.approvalcrest.matcher.AbstractTestMetaBase;
 
-import java.lang.reflect.Method;
 import java.nio.file.Path;
 import java.util.Objects;
 
@@ -31,8 +30,8 @@ public class TestNgTestMeta extends AbstractTestMetaBase {
     private static boolean isTestMethod(StackTraceElement element) {
         try {
             Class<?> clazz = Class.forName(element.getClassName());
-            Method method = findMethod(clazz, element.getMethodName());
-            return method != null && method.isAnnotationPresent(Test.class);
+            return findMethods(clazz, element.getMethodName()).stream()
+                    .anyMatch(method -> method.isAnnotationPresent(Test.class));
         } catch (Throwable e) {
             return false;
         }